Invert the JNI call logic
With this change, instead of calling native code every once in a while to get
the current orientation data, the native code itself calls the relevant listener
whenever new orientation data is available. This may help reduce the latency,
simplify Java logic, and reduce number of threads.

+++ app/src/main/java/com/ginkage/wearmouse/input/MouseSensorListener.java
@@ -91,7 +91,7 @@ public class MouseSensorListener implements SensorService.OrientationListener {
     }
 
     @Override
-    public void onOrientation(float[] quaternion) {
+    public void onOrientation(double[] quaternion) {
         double q1 = quaternion[0]; // X * sin(T/2)
         double q2 = quaternion[1]; // Y * sin(T/2)
         double q3 = quaternion[2]; // Z * sin(T/2)
